requirements.txt
yfinance>=0.2.40
pandas>=2.0
numpy>=1.24
scipy>=1.11
python-dotenv>=1.0
google-genai>=0.8
anthropic>=0.44
twilio>=9.0
alpaca-py>=0.20
feedparser>=6.0
requests>=2.31
